requirements.txt
google-cloud-storage
aiohttp
pyyaml
duckdb
pandas
datetime